Core: port to TypeScript (close #526)
