+++ core/.gitignore
@@ -1,0 +1,4 @@
+index.js
+lib/*.js
+lib/*.js.map
+.tscache
