+++ core/lib/base64.ts
@@ -1,0 +1,85 @@
+/*
+ * Copyright (c) 2013 Kevin van Zonneveld (http://kvz.io)
+ * and Contributors (http://phpjs.org/authors)
+ * 
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to
+ * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
+ * of the Software, and to permit persons to whom the Software is furnished to do
+ * so, subject to the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+// Deprecated and removed in TypeScript
+declare function unescape(s: string): string;
+
+/**
+ * Encode string as base64.
+ * Any type can be passed, but will be stringified
+ *
+ * @param data string to encode
+ * @returns base64-encoded string
+ */
+export function base64encode(data: string): string {
+	// discuss at: http://phpjs.org/functions/base64_encode/
+	// original by: Tyler Akins (http://rumkin.com)
+	// improved by: Bayron Guevara
+	// improved by: Thunder.m
+	// improved by: Kevin van Zonneveld (http://kevin.vanzonneveld.net)
+	// improved by: Kevin van Zonneveld (http://kevin.vanzonneveld.net)
+	// improved by: Rafał Kukawski (http://kukawski.pl)
+	// bugfixed by: Pellentesque Malesuada
+	// example 1: base64_encode('Kevin van Zonneveld');
+	// returns 1: 'S2V2aW4gdmFuIFpvbm5ldmVsZA=='
+	// example 2: base64_encode('a');
+	// returns 2: 'YQ=='
+	// example 3: base64_encode('✓ à la mode');
+	// returns 3: '4pyTIMOgIGxhIG1vZGU='
+
+	var b64 = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/=';
+	var o1, o2, o3, h1, h2, h3, h4, bits, i = 0,
+		ac = 0,
+		enc: string,
+		tmp_arr: Array<string> = [];
+
+	if (!data) {
+		return data;
+	}
+
+	data = unescape(encodeURIComponent(data));
+
+	do {
+		// pack three octets into four hexets
+		o1 = data.charCodeAt(i++);
+		o2 = data.charCodeAt(i++);
+		o3 = data.charCodeAt(i++);
+
+		bits = o1 << 16 | o2 << 8 | o3;
+
+		h1 = bits >> 18 & 0x3f;
+		h2 = bits >> 12 & 0x3f;
+		h3 = bits >> 6 & 0x3f;
+		h4 = bits & 0x3f;
+
+		// use hexets to index into b64, and append result to encoded string
+		tmp_arr[ac++] = b64.charAt(h1) + b64.charAt(h2) + b64.charAt(h3) + b64.charAt(h4);
+	} while (i < data.length);
+
+	enc = tmp_arr.join('');
+
+	var r = data.length % 3;
+
+	return (r ? enc.slice(0, r - 3) : enc) + '==='.slice(r || 3);
+}
+
