+++ core/tests/intern.js
@@ -1,7 +1,7 @@
 /*
  * JavaScript tracker core for Snowplow: tests/intern.js
  * 
- * Copyright (c) 2014 Snowplow Analytics Ltd. All rights reserved.
+ * Copyright (c) 2014-2016 Snowplow Analytics Ltd. All rights reserved.
  *
  * This program is licensed to you under the Apache License Version 2.0,
  * and you may not use this file except in compliance with the Apache License Version 2.0.
@@ -15,41 +15,6 @@
 
 define({
 
-	proxyPort: 9000,
-	proxyUrl: 'http://localhost:9000/',
-
-	capabilities: {
-		'selenium-version': '2.39.0',
-		'build': 'process.env.TRAVIS_JOB_ID'
-	},
-
-	environments: [
-		{ browserName: 'internet explorer', version: '11', platform: 'Windows 8.1' },
-		{ browserName: 'internet explorer', version: '10', platform: 'Windows 8' },
-		{ browserName: 'internet explorer', version: '9', platform: 'Windows 7' },
-		{ browserName: 'firefox', version: '27', platform: [ 'OS X 10.6', 'Windows 7', 'Linux' ] },
-		{ browserName: 'chrome', version: '32', platform: [ 'OS X 10.6', 'Windows 7', 'Linux' ] },
-		{ browserName: 'safari', version: '6', platform: 'OS X 10.8' },
-		{ browserName: 'safari', version: '7', platform: 'OS X 10.9' }
-	],
-
-	maxConcurrency: 3,
-	useSauceConnect: true,
-
-	// Connection information for the remote WebDriver service.
-	webdriver: {
-		host: 'localhost',
-		port: 4444
-	},
-
-	// Configuration options for the module loader; any AMD configuration options supported by the Dojo loader can be
-	// used here
-	loader: {},
-
-	// Functional test suite(s) to run in each browser once non-functional tests are completed
-	functionalSuites: [],
-
-	// A regular expression matching URLs to files that should not be included in code coverage analysis
-	excludeInstrumentation: /^tests\//
+	excludeInstrumentation: /^(?:tests|node_modules)\//
 
 });
