+++ core/tslint.json
@@ -1,0 +1,5 @@
+{
+  "rules": {
+    "indent": [true, "tabs"]
+  }
+}
